Open3DMotion additional tests on event groups, fix to reading of event groups in MOBL, auto-detection and writing of legacy XMove file format (XMove1), XMove2 format information now included in new XMove files by default, examples of current and legacy XMove provided for unit tests.

+++ src/Open3DMotion/OpenORM/Mappings/RichBinary/BinaryFieldSpec.cpp
@@ -10,7 +10,7 @@
 namespace Open3DMotion
 {
 	const char field_type_name_uint8[] = "byte";
-	const char field_type_name_int32[] = "Int32";
+	const char field_type_name_int32[] = "long";
 	const char field_type_name_float32[] = "float";
 	const char field_type_name_float64[] = "double";
